requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.3
cron_descriptor==2.0.6
cryptography==46.0.1
cycler==0.12.1
daphne==4.2.1
decorator==5.2.1
Django==4.2.15
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-timezone-field==7.1
djangorestframework==3.16.1
dnspython==2.8.0
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
dv_processing==2.0.2
et_xmlfile==2.0.0
flatbuffers==25.2.10
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.1
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.1
numba==0.62.1
numpy==2.3.3
onnx==1.18.0
onnxruntime==1.22.1
onnxruntime-gpu==1.22.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==6.31.1
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pyreadline3==3.5.4
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zope.interface==8.0
